# Server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Chord API doc",
        "version": "1.0.1",
        "description": "YourChords - Build a system for instrument players to find, memorize, and practice with the chords of songs",
        "contact": {
            "name": "Yourchord Team Here",
            "email": "huydnse140375@fpt.edu.vn"
        }
    },
    "paths": {
        "/login": {
            "post": {
                "summary": "Đăng nhập người dùng",
                "requestBody": {
                    "description": "Yêu cầu đăng nhập người dùng",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "Tên người dùng"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "Mật khẩu người dùng"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Role": "user",
                                    "Result": {
                                        "username": "nguoidung1",
                                        "role": "user"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi khi gửi yêu cầu",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "Lỗi khi gửi yêu cầu"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sai tên đăng nhập hoặc mật khẩu",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "Sai tên đăng nhập hoặc mật khẩu"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tài khoản đang chờ xét duyệt hoặc bị vô hiệu hóa",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "ErrorPending",
                                    "Error": "Tài khoản đang chờ xét duyệt hoặc bị vô hiệu hóa",
                                    "ban": "Pending"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/requestAccountMusician/{username}": {
            "put": {
                "summary": "Yêu cầu tài khoản âm nhạc",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yêu cầu tài khoản âm nhạc thành công",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi khi gửi yêu cầu",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "Lỗi khi gửi yêu cầu"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/requestAccountChordValidator/{username}": {
            "put": {
                "summary": "Yêu cầu tài khoản xác minh hợp âm",
                "parameters": [
                    {
                        "in": "path",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yêu cầu tài khoản xác minh hợp âm thành công",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi khi gửi yêu cầu",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "Lỗi khi gửi yêu cầu"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/acceptAccountMusician/{username}": {
            "put": {
                "summary": "Accept a musician account",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The username of the musician account to accept"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "update song error in sql"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rejectAccountMusician/{username}": {
            "put": {
                "summary": "Reject a musician account",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The username of the musician account to reject"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "update song error in sql"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/acceptAccountChordValidator/{username}": {
            "put": {
                "summary": "Accept a chord validator account",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The username of the chord validator account to accept"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "update song error in sql"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rejectAccountChordValidator/{username}": {
            "put": {
                "summary": "Reject a chord validator account",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The username of the chord validator account to reject"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "update song error in sql"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/createPlaylist/{username}": {
            "post": {
                "summary": "Create a new playlist",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "collection_name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Playlist created successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Success",
                                    "Result": {}
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "example": {
                                    "Status": "Error",
                                    "Error": "Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}